# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yemen Flow Hub API",
        "description": "Internal regulatory workflow API for CBY import financing requests.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current API server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List users with filters",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Users retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "role": {
                                        "type": "string"
                                    },
                                    "bank_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user details",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "responses": {
                    "200": {
                        "description": "User retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "responses": {
                    "200": {
                        "description": "User updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deactivate user",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "responses": {
                    "200": {
                        "description": "User deactivated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            }
        },
...
        "/api/voting/{importRequest}/vote": {
            "post": {
                "tags": [
                    "Voting"
                ],
                "summary": "Cast executive vote",
                "operationId": "1ca88e5bb69b121f20e19065d9684f4d",
                "parameters": [
                    {
                        "name": "importRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vote"
                                ],
                                "properties": {
                                    "vote": {
                                        "type": "string",
                                        "enum": [
                                            "APPROVE",
                                            "REJECT"
                                        ]
                                    },
                                    "justification": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vote cast"
                    }
                }
            }
        },
        "/api/voting/{importRequest}/director-decide": {
            "post": {
                "tags": [
                    "Voting"
                ],
                "summary": "Director tie-break decision",
                "operationId": "ab589fa29be26f6da15647edf67df9e6",
                "parameters": [
                    {
                        "name": "importRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vote"
                                ],
                                "properties": {
                                    "vote": {
                                        "type": "string",
                                        "enum": [
                                            "APPROVE",
                                            "REJECT"
                                        ]
                                    },
                                    "justification": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Director decision applied"
                    }
                }
            }
        },
        "/api/voting/{importRequest}/override": {
            "post": {
                "tags": [
                    "Voting"
                ],
                "summary": "Executive director override and finalize",
                "operationId": "fec836b6254894ed89dcd82da937c166",
                "parameters": [
                    {
                        "name": "importRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "decision",
                                    "justification"
                                ],
                                "properties": {
                                    "decision": {
                                        "type": "string",
                                        "enum": [
                                            "APPROVE",
                                            "REJECT"
                                        ]
                                    },
                                    "justification": {
                                        "type": "string",
                                        "maxLength": 3000,
                                        "minLength": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Override decision applied"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctumCookie": {
                "type": "apiKey",
                "description": "Sanctum stateful cookie auth for SPA",
                "name": "laravel_session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum personal access token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}